requirements.txt
langchain>=0.0.267
langchain-core>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.4.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.3.0

# Testing dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
